% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asis_chunk.R
\name{asis_chunk}
\alias{asis_chunk}
\title{asischunk code}
\usage{
asis_chunk()
}
\value{
Inserts \code{```{asis}\\n\\n```}
}
\description{
Addin for RStudio for inserting asis chunks. Useful for creating branching documents where markdown code is run conditionally.

See \code{details} for setting up key command.
}
\details{
How to set up key command in RStudio:

After installing package.
Go to:

\code{Tools} >> \code{Addins} >> \code{Browse Addins} >> \code{Keyboard Shortcuts}.

Find \code{"asis code chunk"} and press its field under \code{Shortcut}.

Press desired key command.

Press \code{Apply}.

Press \code{Execute}.
}
\author{
Enrico Manlapig, \email{emanlapig@westmont.edu}
}
